requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==5.1.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pip==23.2.1
pluggy==1.5.0
py==1.11.0
pytest==7.4.4
requests==2.32.3
retry==0.9.2
timeout-decorator==0.5.0
urllib3==2.2.3
